Fix a deadlock bug of gfmd, reported as #274.
When gfmd received SIGTERM, a thread which received the signal in gfmd
took the giant lock first and then tried to lock the journal file mutex.
In contrast, the journal apply thread in gfmd locked the journal file
mutex first, then it tried to take a giant lock.  They caused
the deadlock.