fix #918 - changing gfmd port number by "gfmdhost -m -p" makes gfmd crash
this is due to inconsistency between gfmd and libgfarm
about what is the primary key of gfmd host record.
gfmd thinks that only hostname is the primary key,
but libgfarm thnks that a pair of hostname and port number is the primary key.

make libgfarm think what gfmd does to make them consistent.