merge r10367 from the 2.6 branch:
fix #980 - typo in metadb_server_list setting in gfarm2.conf may cause assertion failure 

it's desired to introduce new error message (to liberror.c) to clarify
that the cause of the error is a typo, but because this error code
may be transfered via gfmd protocol, I chose "invalid argument"
for this error.

this also changes gfmd behavior.
before this change, "no memory" error which happens in mdhost_updated()
was silently ignored, but now it's logged and reported to the corresponding
gfmd client.